Proemmer/transport size bug (#26)
* Set TransportSizeToElementSize always to 1

* fixed some tests and added more targets

+++ dacs7/src/Dacs7Cli/ServeCommand.cs
@@ -41,7 +41,7 @@ namespace Dacs7Cli
                             Port = portOption.HasValue() ? int.Parse(portOption.Value()) : 102,
                             Tags = dataareas.HasValue() ? dataareas.Values : null,
                             DataProvider = dataProvider.HasValue() ? dataProvider.Value() : null,
-                            MaxPduSize = pduSize.HasValue() ? ushort.Parse(pduSize.Value()) : 960
+                            MaxPduSize = pduSize.HasValue() ? ushort.Parse(pduSize.Value()) : (ushort)960
                         }.Configure();
                         var result = await Serve(options, options.LoggerFactory);
